backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
pyodbc==5.2.0
azure-identity==1.21.0
azure-ai-contentsafety==1.0.0
openai==1.63.2
python-dotenv==1.0.1
websockets>=13.0
requests>=2.31.0
